+++ src/items/PostItem.php
@@ -327,6 +327,11 @@ class PostItem
             $content = GeneralHelper::analyzeACF($postItem, $content);
         }
 
+        // Yoast SEO Data
+        if (MigrateFromWordPressPlugin::$plugin->settings->yoastSEO && isset($postItem->yoast_head_json)) {
+            $content['fields']['yoastSEO']['value'] = $postItem->yoast_head_json;
+        }
+
         // Save uri in cache for later
         $contentIds = json_decode(Craft::$app->cache->get('migrate-from-wordpress-pages-posts-id-and-url'), true);
         $contentIds[$postItem->link] = $postItem->id;